requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
numpy==1.21.4
pytz==2021.3
scipy==1.7.2
sqlparse==0.4.2
